% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversedata-package.R
\docType{package}
\name{diversedata}
\alias{diversedata-package}
\alias{diversedata}
\title{A Curated Collection of Datasets on Diverse Topics}
\description{
\code{diversedata} is an R package that offers a growing collection of well-documented,
real-world data sets centered on themes of equity, diversity, and inclusion (EDI).

Each data set includes contextual background and is curated to encourage thoughtful
engagement with social topics through data. Example use cases and vignettes are
included to demonstrate practical applications in R.

This package aims to help learners and educators move beyond generic datasets
by incorporating content that reflects diverse lived experiences and societal issues.

For more information, visit the project website: \url{https://diverse-data-hub.github.io/}
}
\section{Available datasets}{

\itemize{
\item \code{bcindigenousbiz}: Indigenous-owned businesses in British Columbia, Canada.
\item \code{wildfire}: Historical wildfire incidents in Alberta, Canada.
\item \code{globalrights}: Global indicators on LGBTIQ+ rights and related social and economic variables, compiled from Our World in Data.
\item \code{hcmst}: U.S. survey on how couples meet and stay together (2017–2022), including impacts of dating apps and COVID-19. Source: Rosenfeld et al. (2023), Stanford University Libraries.
\item \code{womensmarchmadness}: NCAA Women’s Basketball Tournament stats and outcomes.Source: FiveThirtyEight GitHub, CC BY 4.0.
\item \code{genderassessment}: Company-level gender equality indicators and overall score from the World Benchmarking Alliance's 2024 Gender Assessment.
}
}

\section{Usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(diversedata)
data(bcindigenousbiz)
head(bcindigenousbiz)

data(wildfire)
head(wildfire)
}\if{html}{\out{</div>}}
}

\section{Citation}{

If you use \code{diversedata} in teaching or research, please cite the package:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{citation("diversedata")
}\if{html}{\out{</div>}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/diverse-data-hub/diversedata}
}

}
\author{
\strong{Maintainer}: Katie Burak \email{kburak@stat.ubc.ca} (PhD) [copyright holder]

Authors:
\itemize{
  \item Elham E Khoda (PhD) [copyright holder]
  \item Azin Piran
  \item Francisco Ramirez
  \item Siddarth Subrahmanian
}

Other contributors:
\itemize{
  \item Stephanie Ta \email{stephanietaprofessional@outlook.com} [contributor]
}

}
