% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{bcindigenousbiz}
\alias{bcindigenousbiz}
\title{BC Indigenous Business Directory}
\format{
A data frame of Indigenous-owned businesses in BC
}
\source{
Government of BC
}
\usage{
bcindigenousbiz
}
\description{
BC Indigenous Business Directory
}
\keyword{datasets}
