% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{package}
\name{diversedata-package}
\alias{diversedata-package}
\title{diversedata: A Curated Collection of Datasets on Diverse Topics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides instructors with a convenient resource to incorporate diverse topics in the classroom by offering a curated collection of datasets related to equity, diversity, and inclusion (EDI).
}
\author{
\strong{Maintainer}: Azin Piran \email{az.p.ctc@gmail.com}

Authors:
\itemize{
  \item Francisco Ramirez \email{fraramfra@gmail.com}
  \item Siddarth Subrahmanian \email{sidsub@student.ubc.ca}
  \item Katie Burak (PhD)
  \item Elham E Khoda (PhD)
}

}
\keyword{internal}
