% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{package}
\name{diversedata}
\alias{diversedata}
\title{A Curated Collection of Datasets on Diverse Topics}
\description{
Diverse Data Hub is an open educational resource offering curated datasets
focused on equity, diversity, inclusion, and other socially relevant topics.

It is designed to support students, educators, and researchers in accessing
and working with meaningful data in their teaching, learning, and analysis.

Datasets are available through the \code{diversedata} R package, allowing for straightforward
integration into data science workflows. Each dataset includes detailed documentation
and contextual background to support informed exploration and connection to real-world topics.

Example case studies are also included to illustrate practical applications.
}
\section{Available datasets}{

\itemize{
\item \code{bcindigenousbiz}: Indigenous-owned businesses in British Columbia, Canada.
}
}

\section{Usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(diversedata)
data(bcindigenousbiz)
head(bcindigenousbiz)
}\if{html}{\out{</div>}}
}

\section{Citation}{

If you use \code{diversedata} in teaching or research, please cite the package:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{citation("diversedata")
}\if{html}{\out{</div>}}
}

\seealso{
\link{bcindigenousbiz}
}
\author{
\strong{Maintainer}: Siddarth Subrahmanian \email{sidsub@student.ubc.ca}

Authors:
\itemize{
  \item Azin Piran
  \item Francisco Ramirez
  \item Katie Burak (PhD)
  \item Elham E Khoda (PhD)
}

}
