% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\docType{data}
\name{wildfire}
\alias{wildfire}
\title{Alberta Historical Wildfire Dataset}
\format{
A data frame contains information about historical wildfire incidents in Alberta.
}
\source{
\url{https://open.alberta.ca/licence}
Data retrieved from Alberta Open Data under the Alberta Open Government Licence.
}
\usage{
wildfire
}
\description{
This dataset contains information about historical wildfire incidents in Alberta.
}
\keyword{datasets}
