% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womensmarchmadness.R
\docType{data}
\name{womensmarchmadness}
\alias{womensmarchmadness}
\title{NCAA Women’s March Madness Tournament Dataset}
\format{
A data frame containing game-level and team-level tournament statistics.
}
\source{
\url{https://github.com/fivethirtyeight/data/tree/master/ncaa-womens-basketball-tournament}
Data sourced from FiveThirtyEight under the Creative Commons Attribution 4.0 International License.
}
\usage{
womensmarchmadness
}
\description{
Historical data on the NCAA Women’s Basketball Tournament, including team performance,
scores, seeds, and outcomes.
}
\keyword{datasets}
